Bug fix: don't render extra copies of busy events!
God this took me forever to figure out.
I forgot that addEventListener adds a new event listener every time,
so save button was getting overloaded with tons of redundant listeners
that all did the same thing.
The better fix would probably be to recreate the modal each time,
rather than just reusing the same button, but this works for now